+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.23.0] - 2025-02-07
 ### Added
   - Introdotta funzionalità per svuoltare la cache del Token OAuth
-
+  - Aggiunta possibilità di doppia approvazione del flusso di richiesta straordinari da parte del responsabile
+    di gruppo e del responsabile di sede.
+    
 ### Changed
   - Cambiati defaut scadenza cache token OAuth
   - Corretto salvataggio ExternalId dei gruppi